requirements.txt
flask
flask-cors
requests
pillow
opencv-python
pyzbar
numpy